Bug 2759: Update MIDAS to use new point set distance to plane facility
The default MITK functionality, until recently, was to render
points in 2D at a varying size, depending on the distance from
the point.

For MIDAS, we provided a "show distant points" property, and modified
MITK to remove the above facility, so that if this property was true
points would ONLY show on the given slice. This is useful if there
are many many points.

However, MITK recently provided a property where you can specify
the distance tolerance. This means we can use this property, and
remove the changes from our MITK fork.