Bug #2974: Interactor added to the thumbnail window
A new module, niftkThumbnail has been created because the interaction
definitions (xml) cannot be loaded from plugins, only from modules.

The QmitkThumbnailRenderWindow has been moved to this module. (In the
past it was in the niftkDnDDisplay module, and has been moved to the
thumbnail plugin to separate the thumbnail viewer functionality from
the DnD display.

The interactor does not do anything at the moment.