Bug #2974: Bounding box pan and zoom support
Known issues:

There is currently no way to select the position in the tracked window,
since we have only the renderer and the vtk render window, but not the
multi widget. A single render window, however, does not know the concept
of 'selected point'.

As a consequence, if you start zooming, the focus point will be the point
where you started the interaction, and the tracked window will be zoomed
around that point. But the crosshair will not be moved to that position.
(Since we cannot gain a reference to that.)

If the crosshair is in a StdMultiWidget of the DnDDisplay and the magnification
is bound across its windows, then the selected window will zoom around
the focused point in the thumbnail window, but the other windows of the
multi widget will zoom around the crosshair.