Bug #2974: The focus point issue with the previous commit is fixed
The display geometry modification observer class in the niftkMultiWindowWidget
now calculates the focus point of the zooming, and notifies the multi window
widget if a new zooming operation has started, i.e. the focus point has changed.

The multi window widget then puts the crosshair to the focus point, so that
the zooming starts from the cross hair in each render window (if the
magnification is bound across the windows).