Bug #2625: Focus not stolen when changing window layout
When you changed window in the DnD Display and a MIDAS volume editor
view was switched on, then the layout of its auxiliary view has changed
so that it shows the image always of a different orientation.

However, the SetWindowLayout function also set the focus to one of
the windows of the new layout, what resulted that the focus was
stolen from the DnDDisplay.

Consequently, if you pressed 'a' or 'z' to change a slice then the
slice was changed according to the orientation of the volume editor,
not the DnD display, since the state machine of the volume editor
received the event. (It had got the focus.)

If you wanted to switch orientation according to the selected window
in the DnD Display, you had to click into it again, to put the focus
back.

This commit removes the 'setFocus' calls from the SetWindowLayout
function.