Bug #3029: IsCrossingLine does not calculate "half way" coordinates for contour edges
Earlier the draw tool saved every corner point along the contour, but not the side
points. The region growing filter had a check if a region growing step "crosses"
a contour edge. This algorithm calculated the voxel side coordinates for each
contour edge (each voxel, basically) by the mean of the coordinates.

Since now the voxel side coordinates are stored by the draw tool, there is no need
to calculate the voxel side coordinates again and again.

This commit contains refactoring of the internal logic of ConditionalAddPixel
to make it easier to read. The behavior is equivalent with the previous code.