Bug #2627: Window cursors/magn. binding can controlled together if viewer cursors/magn. are bound
If the cursor positions are bound across the viewers then switching on/off
the window cursor binding affects each viewers.

If the cursor positions are not bound across the viewers than you can
control the window cursor binding independently for the individual viewers,
that is you can bind the cursors within one viewer, but not within another.

The same applies for the magnification binding.