Bug #2627: Do not reinitialise geometries after nodes dropped
The multi viewer widget computes the bounding box for the dropped nodes
and reinitialises the rendering manager with the new bounding box. As a
result of this, the geometry is replaced *again* in the viewers.

If a volume editor is switched on then after dropping a node, if you
move the crosshair upwards in the main window (keeping the left button
pressed and moving the mouse), you see the crosshair in the auxiliary
viewer of the volume editor moving in the opposite direction vertically.

The likely reason is that when you call the mitk::RenderingManager::InitializeViews
then you skip the logic of recomputing the correct geometry for each
slice navigation controllers, as done in niftkMultiWindowWidget::SetGeometry,
and so, the geometry of the renderer and the slice navigation controllers
will be inconsistent.

After dropping nodes the geometry of the multi window widget is set correctly
(probably initiated by the visibility manager), so there is no need to
initialise the views of the rendering manager with a recomputed bounding box
again.