Bug #3146: Minimise render window updates
Render windows should not be updated while the viewer is
in an inconsistent state, i.e. m_BlockSignals is true.
This change updates the render windows when the signals
are unblocked. Only the visible render windows are updated
and only those whose content has changed. (Selected position,
cursor position, scale factor etc.)

Another change is that the focus is put to the selected
render window only when the signals are unblocked.
(And if the selected render window has changed and it is
visible.)