Bug #3191: More thorough test for selecting position by interaction
Note that some failing tests are disabled at the moment.

E.g. when 2x2 window layout is selected and you click in a render
window, three CursorPositionChanged events are emitted even when
only two cursor positions change.

Another problem is that when you switch window layout, the region
should be centred, and the cursor should be in the middle of the
window. However, if you click in the middle of the window, a
SelectedPositionChanged and CursorPositionChanged events are
emitted what shows that the positions had not been in the centre,
actually.