Bug #3237: Do not use extents from display geometry
As a consequence of the redesign of the DnD display module,
the world geometry of the display geometries are not yet initialised
when an image is dragged on the viewer. This is because the SNC
signals are suppressed, and the initialisation will happen when
these suppressed signals are released at BlockUpdate(false).

Therefore, when the image is dropped and we want to find a good
scale factor, we should not rely on the mm extents of the world
geometry of the display geometries, but they should be retrieved
from the world geometry of the viewer.

This commit also contains some clean-up of the 'fitting' logic.