Bug #3250: Package dependencies minimised
The MITK_USE_MODULE macro has been replaced by the mitk_use_modules function,
which allows to specify a set of libraries of a package as dependency, e.g.
you can refer to 'ITK|ITKCommon+ITKIOBase'. With the old macro you could only
specify full packages, like 'ITK'.

With this commit we removed the full package dependencies wherever we could
and replaced them with a list of the specific libraries needed.

The last place where we could not do this is the niftkCore module that still
refers to the ITK package. See the notes in the CMakeLists file of the module.