Bug #3285: Fix for renderer geometry initialisation
If the single viewer was initialised not by an image geometry but
a geometry of a renderer then the cursor was shifted by half voxel.

This was because the origin of the renderer geometry is not in the
corner of the bottom-left-back voxel but it is in the middle of the
bottom left edge of that voxel. Therefore, the orientation of the
input geometry should be considered when calculating the origin of
the new geometry.