Bug #3372: MIDAS volume editors use their own tool manager
Rationale:

Before this change the MIDAS volume editors accessed the tool manager
through the mitk::ToolManagerProvider microservice (singleton) that
returned the same tool manager instance always.

When the data manager selection changed then the volume editors checked
if the newly selected node is a reference image or segmentation or else.
If it was a segmentation then they retrieved the corresponding (derived)
working data nodes from the data storage and set the working data of the
tool manager. That is, both plugins set the working data of the same tool
manager to different set of nodes.