Bug #3283: Positions restored correctly when switching back to a window layout
Note:
Although the positions are now correct, we do a BlockUpdate(false) call
in the middle of restoring the positions. This is wrong and it causes that
unnecessary events are generated, and other views will be notified about
a temporary selected position change that should happen silently.

There are unit tests for this and they will fail at the moment.

The unit tests should be updated also because of the change what positions
are restored.