Bug #3412: Changes in visibility management logic
The GeneralSegmentorView switches on/off the visibility of the selected render
window instead of the global visibility.

The filtering based on the node name has been disabled. Initial node properties are
according to the first greyscale image parent, even for helper nodes.

Feedback contour node of the paintbrush tool and the draw tool treated
specially because it is created at the top level, and therefore it does
not have a greyscale parent.