Bug #3412: Helper nodes not treated differently
This commit also contains some refactorings and clean-up,
and some preliminary code to support 'non-drag-and-drop'
mode when you could use the visibility check boxes in the
data manager to turn on and off visibilities even for nodes
that have not been dragged into the viewer.

This feature is not finished and the preliminary code is
commented out.