Bug #3505: Fix inheritance hierarchy.
I believe the original intention was:

QmitkNiftyViewApplicationPlugin -> QmitkCommonAppsApplicationPlugin -> berry::AbstractUICTKPlugin.

Before this fix, the hierarchy was wrong, due to the fact (if I recall correctly)
that MITK had an incorrect visibility specifier at some point in the distant past.
So, you couldnt correctly inherit.

This change however does change the order of events of the start method.
But I believe this change is in fact the correct order, that should be applied.