Bug #3538: Refactoring related to Modified() calls
Filter setters should check themselves if they need to raise a Modified event or not.
For some filter it was checked externally.

Another change is that the connected component filter is run on the thresholding tab.
This change might be reverted later because it makes the GUI less responsive.
This should be reconsidered later when we manage to run only the necessary filters
at any step. (Not each of them from the start, like now.)