Bug #3574: MIDASTool key press state machine registered as InteractionEventObserver
This state machine is not a Tool and not an interactor. So that it receives the interaction
events, it needs to (multiple) inherit from mitk::InteractionEventObserver and implement
its Notify function in the following way:

void MIDASToolKeyPressStateMachine::Notify(mitk::InteractionEvent* interactionEvent, bool isHandled)
{
  if (!isHandled)
  {
    this->HandleEvent(interactionEvent, NULL);
  }
}

Moreover, the observer must be registered in the module context as a service.

The current commit also loads the definition of the state machine and its
configuration (xml).