Bug #3574: MIDASToolStateMachine.xml renamed to MIDASToolKeyPressStateMachine.xml
The naming was misleading, since the MIDASTool is a statemachine itself, but
the MIDASToolStateMachine.xml was not its definition, but of the
MIDASToolKeyPressStateMachine class.

The state machine of the MIDASTool class was "dummy" before the migration to the
new interaction framework. Now some equivalent empty definition should be provided
for it.

The config file has been renamed accordingly.