Bug #3574: Do not activate tools when the tool selection box is disabled
This caused crash.

If the tool selection box is disabled then the tools are not registered to
the tool manager ( RegisterClient() ). Then if you activate a tool and another
tool was already active, then its interaction event observer service tries to
be unregistered. But since the tools was not registered into the tool manager,
the observer service is still null, and the attempt to unregister it causes crash.