Bug #3574: Empty state machine definition for seed tool
The seed tool got a dummy state machine definition with only
a start state, no transition. The reason for this is that the
seed tool itself does not implement any action, but delegates
the task to a point set interactor. The point set interactor,
however, is a state machine itself, and has its own XML definition.