Bug #3574: MIDASPointSetInteractor renamed to MIDASPointSetDataInteractor
For two reason:

1. It is derived from mitk::PointSetInteractor, not mitk::PointSetDataInteractor.
(Both class still exist, the latter has been introduced with the new interaction
framework.)

2. The old version of the MIDASPointSetInteractor class might need to be
reintroduced because of a missing feature in the DataInteractor version.
The mitk::InteractionPositionEvent does not have setters, and so, you cannot modify
the captured position and delegate the call to the parent. This was possible
with the old point set interactor (mitk::PositionEvent). It would be necessary to
put the seed always in the middle of a voxel, not at arbitrary world position.