Bug #3574: Reverted to old point set interactor
The old point set interactor has been put back in charge, the one that is
not a DataInteractor. This class is a reminiscent of the old interaction
framework of MITK. Currently both classes (mitk::PointSetInteractor and
mitk::PointSetDataInteractor) are provided by MITK, but it can be expected
that the old class will be removed at some point.

It is easy to switch between the two interactors by commenting/uncommenting
a few lines in the MIDASTool and MIDASSeedTool classes.

Two issues with the PointSetDataInteractor:

If you add several point and start removing them, sometimes suddenly each
point disappears. Then you add a few more points, and some of the previously
disappeared points reappear. And so on...

The InteractionPositionEvent does not have setters, and it is not possible
to 'round' the coordinates to voxel centres before passing them down to the
superclass for processing.

There is a third issue that exists with both point set interactors. The last
point of the point set cannot be removed. It is probably removed from the
point set, actually, but the renderer still renders it.