Bug #3507: Refactorings, change in visibility settings logic
The DataNodeAddedVisibilitySetter and the DataNodeVisibilityTracker is now merged,
that makes it possible to set the visibility in the tracked window to a node when
it is added to the data storage. Before now, the mitkDataNodeAddedVisibilitySetter
did not know the track window, so it was only possible a constant visibility
(either always true or always false) to new nodes.