Bug #3507: Side viewer widget should not set the data storage to its viewer
When you set the data storage on a QmitkStdMultiWidget, it passes it down to its
renderers, that recalculate the time geometry based on the globally visible nodes
in the data storage. That is not good for us, since we want to see the nodes that
are visible in the tracked renderer.

The viewer will access the data storage though the rendering manager that is set
through its constructor.