Bug #3507: SetDataStorage functions removed from all viewer classes
There is no interface any more to set the data storage of the viewer.
The internal QmitkStdMultiWidget (and so its renderers) access the
data storage through the rendering manager that is initialised early,
in the constructor.

Providing a setter would allow to set a different data storage then
what the rendering manager uses, that would be inconsistent. Moreover,
when a data storage is set, the renderers get a new geometry based on
the globally visible nodes. This behaviour is undesired.

Note that the data storage of the rendering manager needs to be
set before a viewer is constructed, and it should not change later.