Bug #3507: Public block function removed from DataStorageListener
The function was only used from DataNodeVisibilityTracker, but it does not
need it any more. The visibility tracker sets the visibility only to the
managed renderers, but not the global visiblity or the visibility in the
tracked window. Since the observers are attached only to these latter
visibility properties (global and tracked renderer), setting the visibility
in the managed renderer cannot cause recursion.