Bug #3591: Thumbnail viewer widget does not use geometry of focused renderer
The widget should never refer to the focused renderer, since it cannot
know if the renderer is in the main display or not. It's the responsibility
of the thumbnail viewer view (in the plugin) to listen to the focus events
and set the renderer to track by the thumbnail viewer widget.

The widget should use the reference to the tracked renderer to any of its
operation.