Bug #3616: Side viewer adds its render windows to the rendering manager of the tracked window
This is not instead of using its own rendering manager (as introduced in #3608)
but in addition to it. That means, when a side viewer is tracking a main window,
it is registered to two rendering managers: to its own one and to the one of the
tracked window.

The latter is required so that any time the main window changes, the contents of
the side viewer is immediately updated.