Bug #3538: Pass segmentation image to morpho pipeline
The pipeline should execute one stage at a time.

Currently, the pipeline is executed from the beginning at every stage.
E.g. if you change the parameters of the dilation, the threshoulding and
erosion steps will be executed as well. Therefore, you do not need to pass
the segmentation image as a parameter.

However, if you want to run only the dilation step, you need to pass the
output of the erosion filter as its input.