Bug #3538: New morpho pipeline parameter: start stage
The previous "stage" parameter is kept, it means the end stage, i.e. the
stage where you are at and until which you want to run the pipeline.

The start stage is the stage from which you want to run the pipeline.
Defaults to 0 that means running the pipeline from the beginning. This is
what has been done until now always.