Bug #3538: Pass binary inputs to morpho pipeline
Currently, the morphological segmentation pipeline is always started from the beginning,
i.e. from the thresholding. It is required to be able to run the pipeline from a given
stage. If the start stage is not the thresholding then the input binary image should be
passed to the pipeline. If the start stage is the rethresholding, the thresholding mask
is needed, too.

The pipeline has been updated to consider the start stage and use the additional input
images if the start stage is not the thresholding.

For now, these additional images are always NULL and the start stage is 0 (thresholding).