Bug #3538: Block signals when setting erosions sliders from code
The erosions upper threshold slider should be initialised from the
thresholdint step sliders, but only when we step forward, not back.

The signals should be blocked for that time, since the pipeline will
be updated because of the tab change, anyway. Avoid unnecessary updates.

The commit contains a major simplification of the morpho view GUI logic
as well.