Bug #3538: Refactoring: MITK to ITK casts done earlier
The goal is to check if the casts are necessary at all. Now the images
are always converted to an ITK image and set as inputs of pipeline filters.

This is wrong because the pipeline has to think that its input has changed
and reruns every filter always.

The solution would be to convert the MITK images to ITK only if their
modification time has changed.