Bug #3538: Pipeline manager stores individual pipelines for each segmentation, not one per type
Currently the pipeline manager (PM) stores one pipeline instance per each
image type, e.g. one for unsigned short reference images and another for
float images.

This saves memory because you can reuse the same pipeline if you work with several
float images for instance, or create several segmantation for the same image.

However, this has a drawback on the computation time. The ITK pipeline mechanism
uses the modification time (MTime) of objects to check which parts of the computation
needs to be re-executed. If we use the same pipeline for several images, the
pipeline will need to re-run all the filters until the current stage when you
switch between images.