Bug #3538: morpho pipeline does not lock input images
The commit contains an MITK hash update to get fix about ImageToItk
and read/write accessors.

MITK bug: 17931

The Disconnect() function of the pipeline has been removed. The inputs
should *not* be disconnected after Update(), otherwise the MTime of
the filter will be updated, and next time it will be re-executed, even
if the MITK image is not modified before the two executions and the
same ITK image is set as input before the next run.

The ImageToItk class is used directly to convert the erosion/dilation
masks to ITK. It puts a read lock the MITK image, opposed to a write
lock as earlier, and sets the IgnoreLock option, so that the paintbrush
tool can modify the image between two executions of the pipeline.

The paintbrush tool calls Modified() on the MITK image as well, not
only on the node.

The reference images is converted to ITK using the AccessByItk macros.
(As before.) The macro does not allow to set if we want read or write
lock. Therefore, we need to release the output image. For this, we
duplicate the image and pass a pointer to the copy to the pipeline.