Bug #3691: --dnd option accepts viewer index
The --viewer-rows and --viewer-columns options have been replaced by --viewer-number
that accepts the number of viewer in "<C>" or "<R>x<C>" format where R is the number of rows
and C is the number of columns of the viewers.

E.g.:

  --viewer-number 2x1

The syntax of the --dnd (or --drag-and-drop) option:

  --dnd <viewer>:<node1>[,<node2>]...

where <viewer> can be given in the following format:

  <column>

or

  <row>,<column>

and the rows and columns are numbered from 1.

If you want to drag and drop images to several viewers, you need to apply the --dnd option
several times.