Issue #3845: Added dependency to LegacyIO
The IO framework in MITK has been redesigned to use microservices.
The old IO factory classes become deprecated and are moved to a legacy
module. As a quick and temporary solution, I added the dependency
to this legacy module. This results in warnings at CMake configuration
and compilation time, but the code compiles.

Later we should updated our code to use the microservice based approach.