Bug #3924: Fixed layer property removed after the workbench is up
The property is set during processing the command line arguments,
so that the layer property is not overwritten when the Data Manager
adds the nodes in the data storage to the tree view.

After the workbench is up (and the data manager is started), the
fixedLayer property should be removed so that the user can rearrange
the nodes by drag and drop.