Bug #4093: Obsolete ITK reader removed
The removed reader class checked if the image is nifti or analyze
and if the NIFTK_DRC_ANALYZE environment variable is set, and
delegated the reading to the correct reader.

Now the MITK framework takes care of selecting the right reader
based on the mimetype of the image, and the ITK image io-s can
tell if they can process it or not.

The image io-s are directly wrapped into an MITK reader and
and registered into the microservice based IO facility in the
niftkCoreObjectFactory class.