Bug #4093: ITK Image IOs moved from MITK module to ITK library
Changes in details:

A DRCAnalyzeImageIO class was present in both places, with the exaxt
same code except that they were based on different version of the
Analyze image io. As noted in the file header comments, the DRC
reader must be based on the ITK 3.16.0 version of the Analyze reader,
therefore this version was kept, and moved from the niftkCore module
to the niftkIO library, so that the command line application can use
the same reader.

The newer AnalyzeImageIO was kept, in case somebody needs a newer
version than 3.16.0.

The 3.16.0 version of the analyze reader is made private and is
not exported any more. It's only purpose is to be a base class of
DRCAnalyzeImageIO.