Bug #4121: Added static initialiser to NifTKImageIOFactory.
This factory must be loaded as soon as niftkITK is loaded.
This means that all command line apps will have our
version of the DRC Analyze reader, and our Nifti reader.

This does not affect the GUI, as the new MITK framework is
micro-service based rather than using the ITK Object Factory.