Issue #4338: Fix for setting default build type
The set(CMAKE_BUILD_TYPE "Release" CACHE ...) command does not put the variable in the cache.
This works well for any other variable but this one. (It should either a bug or an undocumented
behaviour of CMake.) The command only puts the variable in the cache if the FORCE option is
given. As a workaround, we only force set the cache variable if it has not been defined. This
lets the user change the value.

This commit also specifies a hint for the CMake UIs which values are accepted.