Issue #4338: Reverting g++ version number fix for Boost
The fix was made for RHEL 6 systems where g++ 4.9 is installed with the
devtoolset-3 package. This package does not create a g++-4.9 symbolic
link to the g++ command, and it is not recognised as a valid toolset
by the boost configuration.

Although the fix fixed this, it broke the configuration on other linux
systems, e.g. Ubuntu 14.04 with gcc 4.8.

As a workaround, on RHEL 6 you should create the symlink on your file system.