Issue #4354: Removing dependency from application plugin to dnddisplay
This change is necesarry so that the application plugin does not depend
on org.mitk.gui.qt.common transitively. The activator of that MITK plugin
assumes that the MITK workbench is already running when the plugin is started.
However, if it is loaded at the same time when the application plugin (because
of the linking dependency) then this cannot be the case, the workbench cannot
be ready yet.