Issue #4435: Rewrite thumbnail viewer logic
The internal logic of the thumbnail viewer has been simplified a bit.
Now the modifications of the tracked renderer are listened to as well,
not just those of the tracked world geometry. This is important so that
the thumbnail render window is updated when the world geometry of the
tracked renderer is *replaced*, not just when it changes.