Issue #4467: Clean-up MITK initial cache options
Plugins do not need to be manually switched off by the initial cache
if they are not built because of the whitelist, anyway. Some obsolete
options removed that have been renamed in MITK. Some others removed
because they are switched off in MITK by default.

The DCMTK_DIR variable is removed from the MITK initial cache because
we have now a fix for the actual MITK bug 19431 (that this was a
workaround for) in our MITK fork.