Issue #4482: VL version update to switch off examples
On Mac there is a compilation error in the VL examples. We do not need the
examples for NifTK, anyway, so I made a quick fix for VL not to build them
if the BUILD_EXAMPLES CMake flag is off. (This should have been like this
already.) The flag is switched off for every external projects.