Issue #4473: Windows test script commented out
It is currently broken, and the test is performed by the build script
anyway.

The test script is commented out so that all the test phase scripts can
pass and the deploy scripts can be tested. The windows unit tests should
be moved from the build script to the test script later.