Issue #4524: Rename helper functions to avoid linker error with VS2013
We got linker error on Windows about the GetEnvironmentVariable function.
It turned out that renaming the function to anything else resolves the error,
which might be occur because there is another function with the same name in
the Windows kernel.

This commit renames the function to GetEnvVar and has some other minor changes
in the API. (Mostly tidy-up.)