Issue #4539: Do not refer to working data from Tool::Deactivated()
mitk::ToolManager::SetWorkingData(std::vector<>) calls mitk::Tool::Activate(-1)
if the passed working data is empty. The Activate function is called
*after* the internal reference to the working data has been set. It
results that the Deactivated() function of a tool derivant cannot access
the working data that was set at its activation. (It will see the empty
vector.)

The MIDAS tools, however, assign an observer to a point set at
activation, which point set can be found based on the working data. The
observer should be removed when the tool is deactivated.

One possible workaround is to call Activate(-1) *always before* setting
an empty working data.

The current solution is saving a reference to the point set so that the
Deactivated() function can remove the observer through that reference
and it does not need to call GetWorkingData().