Issue #4527: Re-run morpho pipeline after paintbrush editing iff necessary
The morphological segmentation pipeline was re-executed when one of the
working nodes of the paintbrush tool has changed, that is the user edited
the segmentation manually. However, the node change event can occur for
various other reasons as well, e.g. visibility change, selection change etc.

To avoid unnecessary updates, the pipeline was reexecuted only when the actual
image (the data contained by the node) has changed. However, there is an
occasion when the pipeline has to be re-executed even when the data has not
changed. This is when an editing session is finished and the user releases
the mouse button after dragging.

This situation could not be discovered using this approach (listening to
node change events). We could update either at too many times (making the
GUI irresponsive) or to few times (not updating the whole image after using
the paintbrush tool.)

This commit drops this approach and introduces a new event in the paintbrush
tool, using which it can send notifications specificaly about the editing
of its working data.