Issue #3899: Region growing works from seed next to contour
Note that this change breaks the unit test. This is because the manual
contour that the unit test creates (as a binary image) is only one pixel
wide, although the draw tool creates two pixel wide contour when it
converts the contour point set into a binary contour image, setting
the pixels on both side of the contour.

The unit test should therefore be corrected.