Issue #3899: Updating unit test to fixed MIDAS region growing algorithm
The unit test created a one voxel wide line in the manual contour image, as in
the past the draw and poly tools have probably done that. However, the tools now
create two voxel wide lines in the image, setting one voxel on each side of the
lines to the 'border value'.

The region growing algorithm also relies on the original contour (vertex list)
that was not provided by the test.

So, the test was stricter than the actualy use case, and it failed.