Issue #4475: Processing CL args delayed by 1ms
The function that processes the command line arguments of the DnD display
was called from the function that creates the GUI widget for the DnD viewer.

Although the viewer was already created, its parent widget was not, and it
was not visible yet either. Therefore, the FitWindow function could not calculate
the right scaling for the image to display.