Issue #4586: KWStyle config fixes
The Header check got a new parameter and does not work without it. (No default value, apparently.)

The Indent check had to be disabled because it gave false errors.

The function name regex check had to be disabled as well because it did not recognise
the 'catch' keyword and the member initialisation list in the constructor definition
before the body.