Issue #4594: ITK access functions decoupled from general segmentor view
Two of the functions were not static because they used a map for caching pipeline objects,
and the map was stored in a data member of the view. The map has been moved out to the
niftkMIDAS module to a singleton class called niftk::GeneralSegmentorPipelineCache.

This will allow moving all the ITK access functions to the module layer.