Issue #4594: Clean-up: unnecessary and dangerous virtual function removed
The 'CreateConnections' function was virtual and it was called from another
base class function that was creating the widgets. It could happen that
the CreateConnections in the base class refers to a widget that has not
been created yet. Similarly, as you are not allowed to call virtual functions
from a constructor.