Issue #4589: Change default level window initialisation mode to 'percentage'
It was 'midas' that is based on the mean and the standard deviation of the pixel
values. The ITK image statistics filter is used to calculate these parameters,
and this filter does not ignore NaN values in the image. Therefore, the level
window upper and lower limit was NaN when the image had NaN pixels what caused
weird errors in MITK.