Issue #4589: Switching to percentage-based level/window initialisation if NaN values
The MIDAS convention uses the standard deviation from the mean of pixel values, that is
NaN if any of the pixel values are NaN. This would result in an invalid level window
that can cause various problems later, e.g. sliders not working correctly. This change
overrides the MIDAS convention based initialisation and uses the mode based on the
percentage of the data range if the image has any NaN values.