Issue #4569: Pull segmentator-independent controller functions to abstract base class
The aim is to move the focus and data node event listening from the
views to the controllers. No the view notifies the controller about
the focus events and data node change or removal. These are independent
of the pure view functionality, and should therefore be moved to the
controllers.