Issue #4569: Refactoring the slice change handler function in general segmentor controller
These are equivalent transformation on the code logic. Return statements have been eliminated
from the function and the duplicated statements from before the previous exit points have been
removed. This simplifies the logic a little bit, but the main aim is to move the code to the
new slice change handling function that gets the orientation as well. Transferring the code
would not be possible if the function has several exit points.