Issue #4569: Slice change handler functions merged
Finally, the last bit of the old handler function has been removed.

The slice change handling is now in one, single function that will make it possible
to simplify the logic, and handling new cases, thanks to the fact that the new function
is notified about both focus and slice changes.