Issue #4569: Initialise seeds for current slice when changing orientation
A new function has been introduced for placing one seed per region on a
given slice. There was a similar function that did this for the whole
volume when the segmentation was started or restarted. However, finding
seeds for the whole volume every time when changing orientation is too
slow.

The commit contains some further tidy up in the slice change logic, and
some other minor style fixes.

[ci skip]