Issue #4569: Initialise seeds only for current slice instead of whole volume
When you started a segmentation or restart an existing segmentation,
seeds were placed into the disjunct regions on every slice along
the axis perpendicular to the current orientation.

When you changed orientation, no seeds were placed on the new slice.

Discovering all the regions of every slice of an orientation at the
beginning is a bit slow, and it seems unnecessary to me. Not discovering
the seeds after switching orientation is definitely wrong, this is
what has been reported as this issue (#4569).

I tried to initialise seeds for the whole volume when switching orientation,
but this was very slow. Moreover, when I switched back to the previous
orientation, many unneeded seeds have been put on the current slice.

This change initialises the seeds only for the current slice, both
when starting a segmentation or when changing orientation. This is
efficient, although I am not sure if it is sufficient as well. I do not
see, though, what was the purpose to place seeds for the whole volume.

[ci skip]