Issue #4569: Do not store slice indices in undo stack object for slice change
First, the slice index is not sufficient, you would also need to store the slice axes
from before and after the change.

Second, the 3D position is already stored that is sufficient to restore the exact
position.

The slice indexes were checked when do-ing or redo-ing the operation, so that
if the slice are equal then nothing is done. However, if the slice are equal
then the slice change should not get onto the undo stack at all, so this would
be a bug. This situation might have occurred with the old way of listening,
but it surely cannot happen now.