Issue #4569: Apply threshold and detect seeds on new slice when switching orientation
A new 'sliceAxis' argument had to be introduced in some functions that took only a slice
index and assumed that they have to work on the current slice. When changing orientation,
we have to do some calculation on the previous slice, that is different orientation, so
these functions behaved incorrectly or crashed.

When changing orientation, the seeds should not be copied from the previous orientation,
otherwise you see several seeds in the same region.