Issue #4604: Use purple colour for manually drawn contours
The region growing filter is leaking where the manual contour and the segmented contour meets.
Since they had the same colour (green), it was really hard to see where the leakage happens
exactly. This commit does not fix the leakage, but makes it at least easy to spot.

Then the user can draw over some contour segments to prevent the leakage.