Issue #4689: Refactorings
The data manager selection is not tracked in the base segmentor controller any more.
First, it is not its responsibility, it is already done by the view class. Second,
it changed the stored selection only when one image was selected, that could cause
that the controller returned a wrong image as selected image when either several or
zero image was selected.

I tidied up a few other things that are not related, just found them on the way.
E.g. more consistent variable naming, passing colours as a single object rather
than three individual floats, removing unused arguments, passing strings as const
references, and images as raw pointers instead of smart pointers where the function
does not need to store them. And so on...