Issue #4689: Ignore data node selection changes if a segmentation is in process
Simultaneous editing of several segmentation is not allowed. This would
introduced to much complexity, and the current code was not designed with
this feature in mind. E.g. you would need to store the selected image in the
undo/redo operations, you would need an new operation about changing working
data and so on.

This commit also contains replacing a typedef-ed type with its original
form. The typedef hides the original type, and I find it rather confusing
that I cannot tell from a variable easily what operations I can do on it.
That would be the advantage of statically typed languages, would not it?